+++ app/[locale]/[[...slug]]/page.tsx
@@ -3,6 +3,7 @@ import { Category } from 'shared/types/product';
 import { PaginatedResult } from 'shared/types/result/PaginatedResult';
 import GASnippet from 'components/headless/GASnippet';
 import { getTranslations } from 'helpers/i18n/get-translations';
+import fetchPageData from 'helpers/server/fetch-page-data';
 import getServerOptions from 'helpers/server/get-server-options';
 import { getLocalizationInfo } from 'project.config';
 import { Providers } from 'providers';
@@ -10,14 +11,14 @@ import { sdk } from 'sdk';
 import { PageProps } from 'types/next';
 import Renderer from 'frontastic/renderer';
 
-export const fetchCache = 'force-no-store';
+export const dynamic = 'force-dynamic';
 
-export async function generateMetadata({ params }: PageProps): Promise<Metadata> {
+export async function generateMetadata({ params, searchParams }: PageProps): Promise<Metadata> {
   const { locale: nextLocale, slug } = params;
 
   sdk.defaultConfigure(nextLocale);
 
-  const response = await sdk.page.getPage({ path: `/${(slug as string[])?.join('/') ?? ''}` });
+  const response = await fetchPageData(slug as string[], searchParams);
 
   if (response.isError) return {};
 
@@ -37,11 +38,7 @@ export default async function Page({ params, searchParams }: PageProps) {
 
   sdk.defaultConfigure(locale);
 
-  const response = await sdk.page.getPage({
-    path: `/${(slug as string[])?.join('/') ?? ''}`,
-    query: searchParams as Record<string, string>,
-    ...getServerOptions(),
-  });
+  const response = await fetchPageData(slug as string[], searchParams);
 
   if (response.isError) return <></>;
 
@@ -49,8 +46,7 @@ export default async function Page({ params, searchParams }: PageProps) {
 
   const categoriesResult = await sdk.callAction<PaginatedResult<Category>>({
     actionName: 'product/queryCategories',
-    payload: { limit: 99 },
-    query: { format: 'tree' },
+    query: { format: 'tree', limit: 99 },
     ...getServerOptions(),
   });